'make check' now runs successfully on a 64-bit Linux machine (openSUSE 13.1).  Notwithstanding this, there are still known deficiencies outstanding as regards subscripting and serialisation/deserialisation.
git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@1472 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/include/CXXR/Browser.hpp
@@ -78,7 +78,7 @@ namespace CXXR {
 	 *
 	 * @return the number of browser levels currently active.
 	 */
-	static unsigned int numberActive()
+	static size_t numberActive()
 	{
 	    return s_browsers.size();
 	}
@@ -113,7 +113,7 @@ namespace CXXR {
 	 *
 	 * @return Pointer to the Browser at level \a i.
 	 */
-	static Browser* fromOutermost(unsigned int i)
+	static Browser* fromOutermost(size_t i)
 	{
 	    return s_browsers.at(i);
 	}
